HIIT-001 rename: 서포터 컨트롤러 힛잇 컨트롤러로 이름 변경
- 서포터 네이밍은 오해의 소지가 있을 수 있다 판단하였다.
